refactoring, moved stats component into general stats component under a stats feature module
