+++ src/app/data/restful.datasource.ts
@@ -3,7 +3,7 @@ import { Injectable, Inject, InjectionToken } from "@angular/core";
 import { environment } from "../../environments/environment";
 import { HttpClient } from "@angular/common/http";
 import { Observable, throwError } from "rxjs";
-import { StatsModel } from "../stats/stats.model";
+import { GeneralStatsModel } from "../stats/generalStats/generalStats.model";
 import { catchError } from "rxjs/operators";
 
 @Injectable()
@@ -15,23 +15,23 @@ export class RestfulDataSource {
         this.url = environment.backendApiUrl;
     }
 
-    getStatsData(): Observable<StatsModel> {
+    getStatsData(): Observable<GeneralStatsModel> {
         // TODO remove this hardcoded part
         console.debug(`getStatsData invoked, sending request to backend: ${this.url}/stats/2`)
-        return this.http.get<StatsModel>(`${this.url}/stats/2`);
+        return this.http.get<GeneralStatsModel>(`${this.url}/stats/2`);
         // TODO make this more generalized
             // .pipe(catchError((error: Response) => throwError(`Error: ${error.status} 
             // ${error.statusText}`)));
     }
 
-    updateStatsData(statsModel: StatsModel): Observable<StatsModel> {
-        return this.http.put<StatsModel>(
+    updateStatsData(statsModel: GeneralStatsModel): Observable<GeneralStatsModel> {
+        return this.http.put<GeneralStatsModel>(
             `${this.url}/stats/${statsModel.id}`, 
             statsModel);
     }
 
-    createStatsData(statsModel: StatsModel): Observable<StatsModel> {
-        return this.http.post<StatsModel>(
+    createStatsData(statsModel: GeneralStatsModel): Observable<GeneralStatsModel> {
+        return this.http.post<GeneralStatsModel>(
             `${this.url}/stats/`, 
             statsModel);
     }
