+++ src/app/stats/generalStats/generalStats.component.html
@@ -1,0 +1,127 @@
+<div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-3 pb-2 mb-3 border-bottom">
+    <h1 class="h2">Stats</h1>
+</div>
+<div class="pt-2 spinner-border spinner-border text-primary" 
+    *ngIf="!statsLoaded"></div>
+
+<div class="pt-2" *ngIf="existingStatsFound">
+    <h3 class="text-muted">
+        General
+        <button class="float-right btn btn-link"
+            data-toggle="modal"
+            data-target="#changeStatsModal">
+            <i class="fa fa-edit"></i>
+        </button>
+    </h3>
+    <!-- <button class="btn btn-link float-right pt-2"
+        data-toggle="modal"
+        data-target="#exampleModal">
+        <i class="fa fa-edit"></i>
+    </button> -->
+    <table class="table table-hover">
+        <thead>
+            <tr>
+                <th>Weight</th>
+                <th>Height</th>
+                <th>Age</th>
+                <th>Bodyfat %</th>
+            </tr>
+        </thead>
+        <tbody>
+            <tr>
+                <td>{{ stats.weight }} {{ stats.weightUnit }}</td>
+                <td>{{ stats.heightFeet }}'{{ stats.heightInch }}</td>
+                <td>{{ stats.age }}</td>
+                <td>{{ stats.bodyfatPercentage }}%</td>
+            </tr>
+            <tr>
+                <td>
+                    <p [ngClass]="getbmiDisplayStringClassmMap()">
+                        <strong class="text-dark">BMI:</strong> {{ bmiDisplayString }}*
+                    </p>
+                </td>
+                <td></td>
+                <td></td>
+                <td></td>
+            </tr>
+        </tbody>
+    </table>
+    <p class="text-muted px-3">
+        <small>
+            * BMI is only a rough estimate for the general population 
+        </small>
+    </p>
+</div>
+<div *ngIf="noStatsFound">
+    <h3 class="text-muted">
+        Enter your general stats
+        <button class="float-right btn btn-link"
+            data-toggle="modal"
+            data-target="#changeStatsModal">
+            <i class="fa fa-edit"></i>
+        </button>
+    </h3>
+</div>
+
+<div class="modal fade" id="changeStatsModal" tabindex="-1" 
+aria-labelledby="changeStatsModalLabel" aria-hidden="true"
+*ngIf="statsLoaded || noStatsFound">
+<div class="modal-dialog">
+  <div class="modal-content">
+    <div class="modal-header">
+      <h5 class="modal-title" id="changeStatsModalLabel">Edit General Stats</h5>
+      <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+        <span aria-hidden="true">&times;</span>
+      </button>
+    </div>
+    <div class="modal-body">
+      <form [formGroup]="statsForm">
+        <div class="form-group">
+            <label>
+                Weight:
+                <input class="form-control" 
+                    type="text" formControlName="weight">
+            </label>
+          </div>
+          <div class="form-group">
+            <label>
+                Unit:
+                <select class="form-control" formControlName="age"
+                    formControlName="weightUnit">
+                    <option>Lb</option>
+                    <option>Kg</option>
+                </select>
+            </label>
+          </div>
+          <div class="form-group">
+            <label>
+                Height (Feet):
+                <input class="form-control" 
+                    type="text" formControlName="heightFeet">
+            </label>
+          </div>
+          <div class="form-group">
+            <label>
+                Height (Inches):
+                <input class="form-control" 
+                    type="text" formControlName="heightInch">
+            </label>
+          </div>
+          <div class="form-group">
+            <label>
+                Age:
+                <input class="form-control" 
+                    type="text" formControlName="age">
+            </label>
+          </div>
+          <div class="modal-footer">
+            <button class="btn btn-link"
+                type="submit" data-dismiss="modal" (click)="onGeneralStatsSubmit()">
+                  Save
+              </button>
+          </div>
+      </form>
+    </div>
+  </div>
+</div>
+</div>
